backend/requirements.txt
fastapi 
uvicorn
pydantic
pyhton-multipart
google-generativeai